pyproject.toml
[project]
name = "ai-prompt-ctf"
version = "0.2.0"
description = ""
authors = [{ name = "Christo Goosen", email = "christogoosen@gmail.com" }]
requires-python = ">=3.11,<3.14"
readme = "README.md"
dependencies = [
    "fastapi==0.115.12",
    "pydantic-settings>=2.0.3,<3",
    "jinja2>=3.1.5,<4",
    "python-multipart>=0.0.18,<0.0.19",
    "python-dotenv>=1.1.0,<2",
    "clean-text>=0.6.0,<0.7",
    "unidecode>=1.3.7,<2",
    "openai>=1.1.1,<2",
    "bandit>=1.7.5,<2",
    "gunicorn==23.0.0",
    "chromadb>=0.5.7,<0.6",
    "jinja-markdown~=1.210911",
    "slowapi>=0.1.9,<0.2",
    "certifi>=2025.1.31,<2026",
    "transformers>=4.51.0,<5",
    "nest-asyncio>=1.6.0,<2",
    "torch>=2.6.0,<3",
    "accelerate>=1.6.0,<2",
    "mem0ai>=0.1.86,<0.2",
    "ollama>=0.4.7,<0.5",
    "google-adk[eval, litellm]>=0.3.0,<0.4",
    "litellm>=1.67.4.post1,<2",
]

[dependency-groups]
dev = [
    "fastapi-cli>=0.0.7,<0.0.8",
    "pytest>=8.3.5,<9",
    "importanize>=0.7.0,<0.8",
    "memory-profiler>=0.61.0,<0.62",
    "black[jupyter]>=25.1.0,<26",
    "flake8>=6.1.0,<7",
    "flake8-bugbear>=23.12.2,<24",
    "bandit>=1.8.3,<2",
    "pytest-cov>=4.1.0,<5",
    "uvicorn>=0.34.0,<0.35",
    "setuptools>=70.0.0,<71",
    "huggingface-hub==0.30.1",
    "pytest-asyncio>=0.24.0,<0.25",
    "djlint>=1.36.1,<2",
    "anyio>=4.6.2.post1,<5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ctf"]

[tool.black]
line-length = 80
target-version = ["py312"]

[tool.djlint]
exclude=".venv,venv,.tox,.eggs,..."
ignore="H006,H013,H021"



